don't store geo coords in metadata. Store it in street-geo
