+++ src/json-utils_1.1.js
@@ -35,11 +35,6 @@ const getCurrentSceneTitle = () => {
 };
 STREET.utils.getCurrentSceneTitle = getCurrentSceneTitle;
 
-const getGeoData = () => {
-  const currentGeoCoord = AFRAME.scenes[0].getAttribute('metadata').coord;
-  return currentGeoCoord;
-};
-
 /*
 Takes one or more elements (from a DOM queryselector call)
 and returns a Javascript object
@@ -49,7 +44,6 @@ function convertDOMElToObject(entity) {
   const environmentElement = document.querySelector('#environment');
   const referenceEntities = document.querySelector('#reference-layers');
   const sceneEntities = [entity, environmentElement, referenceEntities];
-  const geoData = getGeoData();
 
   // get assets url address
   assetsUrl = document.querySelector('street-assets').getAttribute('url');
@@ -64,7 +58,6 @@ function convertDOMElToObject(entity) {
   return {
     title: 'scene',
     version: '1.0',
-    geoData: geoData,
     data: data
   };
 }
@@ -677,11 +670,6 @@ function createElementsFromJSON(streetJSON) {
   if (sceneTitle) {
     console.log('sceneTitle from createElementsFromJSON', sceneTitle);
     AFRAME.scenes[0].setAttribute('metadata', 'sceneTitle', sceneTitle);
-  }
-
-  const geoData = streetObject['geoData'];
-  if (geoData) {
-    AFRAME.scenes[0].setAttribute('metadata', 'coord', geoData);
   }
 
   const streetContainerEl = document.getElementById('street-container');
